requirements.txt
# Web server
fastapi>=0.110
uvicorn>=0.29

# Computer vision
opencv-contrib-python>=4.9
mediapipe>=0.10.0

# AWS
boto3>=1.34

# Utilities
numpy>=1.24,<2.0
Pillow>=10.0
python-dotenv>=1.0
requests>=2.31
tqdm>=4.66

# Chess logic
python-chess>=1.10

# ML — piece classifier (install torch CPU separately, see Dockerfile)
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.0
torchvision>=0.15
